refactor(manageProject): update folder, file names
